Move permissions. Clean up unneeded SpecimenMigrationService references. Create manage study registration mechanisms for additional views and use for specimen panels.
