+++ specimen/src/org/labkey/specimen/actions/ShowGroupMembersAction.java
@@ -31,7 +31,7 @@ import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
-import org.labkey.api.specimen.security.permissions.ManageSpecimenActorsPermission;
+import org.labkey.specimen.security.permissions.ManageSpecimenActorsPermission;
 import org.labkey.api.study.StudyUrls;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
