+++ specimen/src/org/labkey/specimen/view/SpecimenWebPartFactory.java
@@ -2,7 +2,6 @@ package org.labkey.specimen.view;
 
 import org.jetbrains.annotations.NotNull;
 import org.labkey.api.security.permissions.ReadPermission;
-import org.labkey.api.specimen.view.SpecimenWebPart;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.view.DefaultWebPartFactory;
