+++ specimen/src/org/labkey/specimen/view/chooseImporter.jsp
@@ -9,7 +9,6 @@
 <%@ page import="org.labkey.api.data.Container" %>
 <%@ page import="org.labkey.api.module.ModuleLoader" %>
 <%@ page import="org.labkey.api.security.User" %>
-<%@ page import="org.labkey.api.specimen.SpecimenMigrationService" %>
 <%@ page import="org.labkey.api.study.SpecimenService" %>
 <%@ page import="org.labkey.api.study.SpecimenTransform" %>
 <%@ page import="org.labkey.api.study.StudyUrls" %>
@@ -17,8 +16,9 @@
 <%@ page import="org.labkey.api.util.HtmlString" %>
 <%@ page import="org.labkey.api.util.URLHelper" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
-<%@ page import="static org.labkey.api.util.HtmlString.NBSP" %>
+<%@ page import="org.labkey.specimen.actions.ShowUploadSpecimensAction" %>
 <%@ page import="java.util.Collection" %>
+<%@ page import="static org.labkey.api.util.HtmlString.NBSP" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%
@@ -37,7 +37,7 @@
     HtmlString manageFoldersLink = h(urlProvider(AdminUrls.class).getFolderTypeURL(c));
     HtmlString labkeyEditionsLink = h("https://www.labkey.com/products-services/labkey-server/labkey-server-editions-feature-comparison/");
     HtmlString contactUsLink = h("https://www.labkey.com/about/contact/");
-    HtmlString manuallyImportSpecimensLink = h(SpecimenMigrationService.get().getUploadSpecimensURL(c));
+    HtmlString manuallyImportSpecimensLink = h(urlFor(ShowUploadSpecimensAction.class));
 %>
 
 <style type="text/css">
