+++ study/api-src/org/labkey/api/study/StudyInternalService.java
@@ -2,6 +2,8 @@ package org.labkey.api.study;
 
 import org.labkey.api.annotations.Migrate;
 import org.labkey.api.data.Container;
+import org.labkey.api.security.SecurityManager;
+import org.labkey.api.security.SecurityManager.ViewFactory;
 import org.labkey.api.security.User;
 import org.labkey.api.services.ServiceRegistry;
 import org.labkey.api.specimen.query.SpecimenQueryView;
@@ -48,4 +50,6 @@ public interface StudyInternalService
     Integer getLastSpecimenRequest(Study study);
 
     SpecimenQueryView getSpecimenQueryView(ViewContext context, boolean showVials, boolean forExport, ParticipantDataset[] cachedFilterData, SpecimenQueryView.Mode viewMode, CohortFilter cohortFilter);
+
+    void registerManageStudyViewFactory(ViewFactory factory);
 }
